Adding a checker if user has access to team before change products cat

+++ src/App/Controllers/ProductCategory.ts
@@ -2,6 +2,8 @@ import { Request, Response } from 'express';
 import { getRepository } from 'typeorm';
 import * as Yup from 'yup';
 
+import { checkIfUserHasAccessToTeam } from '../../Functions/Security/UserAccessTeam';
+
 import { Category } from '../Models/Category';
 import { Product } from '../Models/Product';
 import ProductCategory from '../Models/ProductCategory';
@@ -26,6 +28,34 @@ class ProductCategoryController {
             const { id } = req.params;
             const { product_id } = req.body;
 
+            const productRepository = getRepository(Product);
+            const categoryRepository = getRepository(Category);
+
+            const product = await productRepository.findOne({
+                where: { id: product_id },
+            });
+            const category = await categoryRepository.findOne({
+                where: { id },
+                relations: ['team'],
+            });
+
+            if (!product || !category) {
+                return res
+                    .status(400)
+                    .json({ error: 'Category or Product was not found' });
+            }
+
+            const userHasAccess = await checkIfUserHasAccessToTeam({
+                team_id: category.team.id,
+                user_id: req.userId,
+            });
+
+            if (!userHasAccess) {
+                return res
+                    .status(401)
+                    .json({ error: 'You dont have authorization to do this' });
+            }
+
             const repository = getRepository(ProductCategory);
 
             const alreadyExists = await repository.findOne({
@@ -43,22 +73,6 @@ class ProductCategoryController {
                 return res
                     .status(400)
                     .json({ error: 'Product is already in category' });
-            }
-
-            const productRepository = getRepository(Product);
-            const categoryRepository = getRepository(Category);
-
-            const product = await productRepository.findOne({
-                where: { id: product_id },
-            });
-            const category = await categoryRepository.findOne({
-                where: { id },
-            });
-
-            if (!product || !category) {
-                return res
-                    .status(400)
-                    .json({ error: 'Category or Product was not found' });
             }
 
             const productCategory = new ProductCategory();
@@ -94,24 +108,33 @@ class ProductCategoryController {
 
             const repository = getRepository(ProductCategory);
 
-            const exists = await repository.findOne({
-                where: {
-                    category: {
-                        id,
-                    },
-                    product: {
-                        id: product_id,
-                    },
-                },
-            });
+            const exists = await repository
+                .createQueryBuilder('prod_cat')
+                .leftJoinAndSelect('prod_cat.category', 'category')
+                .leftJoinAndSelect('prod_cat.product', 'product')
+                .leftJoinAndSelect('product.team', 'team')
+                .leftJoinAndSelect('team.team', 'temObj')
+                .where('product.id = :product_id', { product_id })
+                .andWhere('category.id = :category_id', { category_id: id })
+                .getOne();
 
             if (!exists) {
                 return res
                     .status(400)
                     .json({ error: 'Product was not in category' });
             }
+            const userHasAccess = await checkIfUserHasAccessToTeam({
+                team_id: exists.product.team[0].team.id,
+                user_id: req.userId,
+            });
 
-            const savedProductCategory = await repository.remove(exists);
+            if (!userHasAccess) {
+                return res
+                    .status(401)
+                    .json({ error: 'You dont have authorization to do this' });
+            }
+
+            await repository.remove(exists);
 
             return res
                 .status(200)
